requirements.txt
colour==0.1.2
PyOpenGL==3.1.0
wheel==0.24.0
